scripts/requirements.portfolio_loader.txt
aiohttp
pathlib
python-dotenv>=0.19.0
